% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_collapse.R
\name{overlap_collapse}
\alias{overlap_collapse}
\alias{overlap_collapse.data.frame}
\alias{overlap_collapse.Date}
\alias{overlap_collapse.POSIXt}
\alias{overlap_collapse.numeric}
\title{Collapsing overlapped ranges}
\usage{
overlap_collapse(x, ...)

\method{overlap_collapse}{data.frame}(
  .data,
  range_cols,
  method = c("outer", "inner", "left", "right")
)

\method{overlap_collapse}{Date}(start, end, method = c("outer", "inner", "left", "right"))

\method{overlap_collapse}{POSIXt}(start, end, method = c("outer", "inner", "left", "right"))

\method{overlap_collapse}{numeric}(x, y, method = c("outer", "inner", "left", "right"))
}
\arguments{
\item{x}{a numeric vector specifying the one border of a range}

\item{.data}{a data.frame, tbl, or grouped_df. If grouped, collapsing will be performed by group.}

\item{range_cols}{unquoted column names that specify the beginning and stopping of the range}

\item{method}{a string, either:

- outer: the collapsed range is the widest one
- inner: the collapsed range is the closest one
- left: the collapsed range is skew to the left
- right: the collapsed range is skew to the right}

\item{start}{start of a period}

\item{end}{end of a period}

\item{y}{a numeric vector specifying the other border of a range}
}
\value{
For data.frame method: a tbl or a grouped_df

For Date/POSIXct/POSIXlt method: a vector of the same class

For numeric method: a numeric vector
}
\description{
A function to collapse different ranges
}
